Move branches and pulls count to github section
